Line count reduction :\
