+++ fuel-derive/README.md
@@ -5,4 +5,4 @@
 [![docs](https://docs.rs/fuel-derive/badge.svg)](https://docs.rs/fuel-derive/)
 [![discord](https://img.shields.io/badge/chat%20on-discord-orange?&logo=discord&logoColor=ffffff&color=7389D8&labelColor=6A7EC2)](https://discord.gg/xfpK4Pe)
 
-This crate contains derive macros for canonical serialization and deserialization. This is used with [`fuel-types/src/canonical.rs`](fuel-types/src/canonical.rs) module which contains the associated traits and their implementations for native Rust types.
+This crate contains derive macros for canonical serialization and deserialization. This is used with [`fuel-types/src/canonical.rs`](fuel-types/src/canonical.rs) module which contains the associated traits and their implementations for native Rust types. It also contains compaction macro `Compressed`, exporeted by `fuel-compression`.
